Alterações no DAO e Servlet
