--- ProjetoJSP/src/main/java/model/DBConnection.java
@@ -1,20 +1,0 @@
-package model;
-
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
-
-public class DBConnection {
-    private static final String URL = "jdbc:mysql://localhost:3306/seu_banco";
-    private static final String USER = "seu_usuario";
-    private static final String PASSWORD = "sua_senha";
-
-    public static Connection getConnection() {
-        try {
-            Class.forName("com.mysql.cj.jdbc.Driver"); // Carrega o driver
-            return DriverManager.getConnection(URL, USER, PASSWORD);
-        } catch (ClassNotFoundException | SQLException e) {
-            throw new RuntimeException("Erro na conexão com o banco de dados", e);
-        }
-    }
-}
