--- ProjetoJSP/src/main/java/model/Pizza.java
@@ -1,13 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package model;
-
-/**
- *
- * @author pedro
- */
-public class Pizza {
-    
-}
